requirements.txt
anthropic
pandas
python-dotenv
matplotlib
